luci: show wifi vlan in the associated wireless stations list
This patch adds a badge to the associated wireless stations with the vlan id
and name. It is displayed in the same color as the network, to which it is
bridged, so the color corresponds to the color in the network configuration
page.

Signed-off-by: Tobias Waldvogel <tobias.waldvogel@gmail.com>
(cherry picked from commit 29fe3f5fdad6bc7c72799adff22847a6e257f7c0)