luci-mod-status: add log filtering to syslog tab
Facility and severity filtering are based on a simple includes() search.
As such, false positives are possible. Although for the majority of
cases, this is still useful. Filtering using logread -z/Z is possible,
but not if a static log-file is configured. logread does not yet handle
severity either. 'not' checkboxes for each invert the respective search
filter.

A raw-text filter is also included as a bonus, whose meaning can be
inverted via the 'not' checkbox.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>
(cherry picked from commit 16beffda15008bd9cb13ae051d43c281746e2d2e)